ieee802154/cc2420: Insert the length of the MPDU as first byte into TXFIFO
The cc2420 takes care about inserting the preamble sequence and the SFD in
hardware. The first byte you need to insert into the FIFO is is the length byte
for the MPDU. Without this the first byte of FCF was used as length. This
results in wrong information about length FCF and more. Thefore the whole
package looks destroyed on the air.

With this patch packets sniffed from the air and decoded with wireshark are
having the expected content and FCS.

Signed-off-by: Stefan Schmidt <stefan@datenfreihafen.org>