ieee802154/cc2420: Fix packet length value.
This is an addition to the last fix for adding the length byte. With hardware
FCS enabled we are obliged to set the length including the 2 bytes for the
FCS. Our skb does only have the len for the data it contains so we need to
adjust it for the length byte.

Signed-off-by: Stefan Schmidt <stefan@datenfreihafen.org>