HACK: ieee802154/cc2420: Avoid gpio_request for fifop.
When gpio_request()'ing the fifop GPIO we get back an -EBUSY. It seems to be
related to fifop being 0 and the gpiolib may have some special handling for 0.
As this is "only" a resource allocation call which does not do anything with the
GPIO we are still able to use it correctly without it and it works as expected.

However this needs to be worked out completely and a real fix is needed.